requirements.txt
cupy_cuda13x==13.6.0
matplotlib==3.10.7
numpy==2.3.4
